refactor(derive): move derive(FromVariant) into its own file
